Change how fits are made
Fits are now made logarithmically, so they should converge to 1 at 0
rather than -Inf. Additionally, a constraint was added so that the lower
limit (y axis) can only be 0, not -Inf. However, this does not assume
that the lower limit IS 0. Additionally, IC values are plotted at their
correct y-values. An IC50 will not necessarily be at 0.5 if there is a
high 'floor' to the curve. While the value reflected this, the place it
was plotted did not, which was confusing.